requirements.txt
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
flake8==4.0.1
greenlet==1.1.2
idna==3.3
mccabe==0.6.1
msgpack==1.0.3
mypy-extensions==0.4.3
neovim==0.3.1
pathspec==0.9.0
platformdirs==2.5.1
py-console==0.1.4
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
pynvim==0.4.3
requests==2.27.1
toml==0.10.2
tomli==2.0.1
tomli-w==1.0.0
typing-extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
